generate changes and search filtering

+++ db/setup.sh
@@ -8,7 +8,10 @@ cd $mybase
 . ../backend/.flaskenv
 dbname=$DB_NAME
 
-echo $dbname
+if psql amazoff -c "SELECT pid FROM pg_stat_activity WHERE state = 'idle'" | grep -q 1; then
+    echo 'ERROR: Close all database connections and run again'
+    exit 1
+fi
 if [[ -n `psql -lqt | cut -d \| -f 1 | grep -w "$dbname"` ]]; then
     dropdb $dbname
 fi
